refactoring - agent file seperation
